+++ backend/index/h3_btree/knn.go
@@ -6,14 +6,14 @@ import (
 	"time"
 )
 
-func (r *CollectionGeohash) KNNTimed(origin geo.Point, n uint64) (geo.Points, time.Duration) {
+func (r *CollectionGeohash) KNNTimed(origin geo.Point, n int) (geo.Points, time.Duration) {
 	t0 := time.Now()
 
 	originHash := r.hash(origin)
 	var points geo.Points
 	for neighbors := range h3_utils.GridDiskInf(originHash) {
 		points = append(points, r.getMany(neighbors)...)
-		if len(points) >= int(n) {
+		if len(points) >= n {
 			break
 		}
 	}
@@ -23,5 +23,5 @@ func (r *CollectionGeohash) KNNTimed(origin geo.Point, n uint64) (geo.Points, ti
 	upperRight := origin.AddLatitude(mostDistance).AddLongitude(mostDistance)
 	bboxedPoints := r.bbox(bottomLeft, upperRight)
 
-	return bboxedPoints.GetClosestViaSort(origin, int(n)), time.Since(t0)
+	return bboxedPoints.GetClosestViaSort(origin, n), time.Since(t0)
 }
