+++ backend/index/kdtree/addapter.go
@@ -74,9 +74,9 @@ func (r *CollectionKDTree) BBoxTimed(bottomLeft geo.Point, upperRight geo.Point)
 	return toConcrete(res), dur
 }
 
-func (r *CollectionKDTree) KNNTimed(origin geo.Point, n uint64) (geo.Points, time.Duration) {
+func (r *CollectionKDTree) KNNTimed(origin geo.Point, n int) (geo.Points, time.Duration) {
 	t0 := time.Now()
-	res := r.impl.KNN(&origin, int(n))
+	res := r.impl.KNN(&origin, n)
 	dur := time.Since(t0)
 
 	return toConcrete(res), dur
